Create 09Partition Equal Subset Sum
Partition Equal Subset Sum
You are given an array arr of N positive integers. Your task is to find if we can partition the given array into two subsets such that the sum of elements in both subsets is equal.

For example, let’s say the given array is [2, 3, 3, 3, 4, 5], then the array can be partitioned as [2, 3, 5], and [3, 3, 4] with equal sum 10.

Input Format
The first line of input contains an integer T representing the number of test cases or queries to be processed.

Then the test case follows.

The first line of each test case contains an integer N, denoting the size of the array.

The second line of each test case contains N single space-separated integers representing the elements of arr.

Output Format
For each test case, print “true” or “false” denoting whether we can partition into two equal subset-sum in a separate line.

Example 1
Input

2
6
3 1 1 2 2 1
5
5 6 5 11 6
Output

true
false
Explanation

For the first test case, the array can be partitioned as ([2,1,1,1] and [3, 2]) or ([2,2,1], and [1,1,3]) with sum 5.

For the second test case, the array can’t be partitioned.