Create 10Value Equal To The Index Value
Value Equal To The Index Value
Ankit and his friend are playing a game in which his friend selects an integer 'N'. He then picks 'N' random numbers and places them in an array/list ‘NUMARRAY’.

He asks Ankit to find all those numbers which are equal to their index values i.e. 'NUMARRAY[i]' = ‘i’ where ‘i’ ranges from 0 to N - 1.

Can you help Ankit find all such numbers?

For example:

Let ‘NUMARRAY’ = [-4, -2, 2, 5]. Here, only 'NUMARRAY[2]' = 2. So, our answer is 2.

Note: If there is no such number present in 'NUMARRAY,’ which equals its index value, then return -1.

Input Format
The first line contains a single integer ‘N’.

The next line contains ‘N’ single space-separated integers denoting the numbers in ‘NUMARRAY.’.

Output Format
Print all the numbers which satisfy the given condition in sorted order.

Example 1
Input:
4
-4 -2 2 5

Output:
2
Explanation:

For ‘i’ = ‘0’, 'NUMARRAY[0]' = -4 which is not equal to index ‘i’, So we don’t include this number in our answer.

For ‘i’ = ‘1’ 'NUMARRAY[1]' = -2 which is not equal to index ‘i’, So we don’t include this number in our answer.

For ‘i’ = ‘2’ 'NUMARRAY[2]' = 2 which is equal to index ‘i’, So we include this number in our answer.

For ‘i’ = ‘3’ 'NUMARRAY[3]' = 5 which is not equal to index ‘i’, So we don’t include this number in our answer.